ui_part: add FSM for album art transmission to fix fast-folder-change corruption
Replaced ad-hoc album art state variables with a 3-state FSM
(ALBUM_IDLE / ALBUM_WAIT_HEADER_ACK / ALBUM_TRANSFERRING):

- Aborting in-flight transfer on folder change: handle_set_folder now
  resets albumState=ALBUM_IDLE immediately, so any chunk responses still
  in the UART pipeline are discarded by state-check guards.

- Out-of-order chunk detection: rpc_album_on_chunk compares
  chunk.offset against nextBytes; mismatches (pipeline remnants from a
  previously aborted transfer) trigger a re-request from the current
  offset rather than corrupting album_compressed.

- 3-second inactivity timeout (album_fsm_tick) resets the FSM if a
  chunk response is lost, preventing the transfer from hanging forever.

- Removed the unconditional rpc_album_end() call that fired after every
  chunk and caused LVGL to re-render partial (corrupt) album art on
  every packet.

- Fixed rpc_request_album_chunk declaration uint16_t->uint32_t to match
  ImageChunk_t field widths and the implementation.

- Removed the now-superseded imageSeq discriminator (was set to g_seq
  after the header send, so it incorrectly pointed one past the header
  packet's own sequence number).

+++ ui_part/player_rpc.h
@@ -135,7 +135,7 @@ void rpc_set_new_folder(int8_t dir);
 void rpc_set_new_track(int8_t dir);
 
 void rpc_set_album_begin(void);
-void rpc_request_album_chunk(uint16_t offset, uint16_t len);
+void rpc_request_album_chunk(uint32_t offset, uint32_t len);
 void rpc_set_album_end(void);
 
 void rpc_start_bt_scan(void);
